Consolidate tests to tsdb package
