+++ tsdb/index/tsi1/index_test.go
@@ -9,10 +9,8 @@ import (
 	"regexp"
 	"testing"
 
-	"github.com/influxdata/influxdb/internal"
 	"github.com/influxdata/influxdb/models"
 	"github.com/influxdata/influxdb/tsdb/index/tsi1"
-	"github.com/influxdata/influxql"
 )
 
 // Bloom filter settings used in tests.
@@ -119,116 +117,6 @@ func TestIndex_MeasurementExists(t *testing.T) {
 		} else if v {
 			t.Fatal("expected measurement to be deleted")
 		}
-	})
-}
-
-// Ensure index can return a list of matching measurements.
-func TestIndex_MeasurementNamesByExpr(t *testing.T) {
-	idx := MustOpenIndex()
-	defer idx.Close()
-
-	// Add series to index.
-	if err := idx.CreateSeriesSliceIfNotExists([]Series{
-		{Name: []byte("cpu"), Tags: models.NewTags(map[string]string{"region": "east"})},
-		{Name: []byte("cpu"), Tags: models.NewTags(map[string]string{"region": "west"})},
-		{Name: []byte("disk"), Tags: models.NewTags(map[string]string{"region": "north"})},
-		{Name: []byte("mem"), Tags: models.NewTags(map[string]string{"region": "west", "country": "us"})},
-	}); err != nil {
-		t.Fatal(err)
-	}
-
-	// Retrieve measurements by expression
-	idx.Run(t, func(t *testing.T) {
-		t.Run("EQ", func(t *testing.T) {
-			names, err := idx.MeasurementNamesByExpr(nil, influxql.MustParseExpr(`region = 'west'`))
-			if err != nil {
-				t.Fatal(err)
-			} else if !reflect.DeepEqual(names, [][]byte{[]byte("cpu"), []byte("mem")}) {
-				t.Fatalf("unexpected names: %v", names)
-			}
-		})
-
-		t.Run("NEQ", func(t *testing.T) {
-			names, err := idx.MeasurementNamesByExpr(nil, influxql.MustParseExpr(`region != 'east'`))
-			if err != nil {
-				t.Fatal(err)
-			} else if !reflect.DeepEqual(names, [][]byte{[]byte("disk"), []byte("mem")}) {
-				t.Fatalf("unexpected names: %v", names)
-			}
-		})
-
-		t.Run("EQREGEX", func(t *testing.T) {
-			names, err := idx.MeasurementNamesByExpr(nil, influxql.MustParseExpr(`region =~ /east|west/`))
-			if err != nil {
-				t.Fatal(err)
-			} else if !reflect.DeepEqual(names, [][]byte{[]byte("cpu"), []byte("mem")}) {
-				t.Fatalf("unexpected names: %v", names)
-			}
-		})
-
-		t.Run("NEQREGEX", func(t *testing.T) {
-			names, err := idx.MeasurementNamesByExpr(nil, influxql.MustParseExpr(`country !~ /^u/`))
-			if err != nil {
-				t.Fatal(err)
-			} else if !reflect.DeepEqual(names, [][]byte{[]byte("cpu"), []byte("disk")}) {
-				t.Fatalf("unexpected names: %v", names)
-			}
-		})
-	})
-}
-
-func TestIndex_MeasurementNamesByExpr_Auth(t *testing.T) {
-	idx := MustOpenIndex()
-	defer idx.Close()
-
-	// Add series to index.
-	if err := idx.CreateSeriesSliceIfNotExists([]Series{
-		{Name: []byte("cpu"), Tags: models.NewTags(map[string]string{"region": "east"})},
-		{Name: []byte("cpu"), Tags: models.NewTags(map[string]string{"region": "west", "secret": "foo"})},
-		{Name: []byte("disk"), Tags: models.NewTags(map[string]string{"secret": "foo"})},
-		{Name: []byte("mem"), Tags: models.NewTags(map[string]string{"region": "west", "country": "us"})},
-	}); err != nil {
-		t.Fatal(err)
-	}
-
-	authorizer := &internal.AuthorizerMock{
-		AuthorizeSeriesReadFn: func(database string, measurement []byte, tags models.Tags) bool {
-			if tags.GetString("secret") != "" {
-				t.Logf("Rejecting series db=%s, m=%s, tags=%v", database, measurement, tags)
-				return false
-			}
-			return true
-		},
-	}
-
-	// Retrieve measurements by expression
-	idx.Run(t, func(t *testing.T) {
-		t.Run("No Filter", func(t *testing.T) {
-			names, err := idx.MeasurementNamesByExpr(authorizer, nil)
-			if err != nil {
-				t.Fatal(err)
-			} else if !reflect.DeepEqual(names, [][]byte{[]byte("cpu"), []byte("mem")}) {
-				t.Fatalf("unexpected names: %v", BytesToStrings(names))
-			}
-		})
-
-		t.Run("EQ", func(t *testing.T) {
-			names, err := idx.MeasurementNamesByExpr(authorizer, influxql.MustParseExpr(`region = 'west'`))
-			if err != nil {
-				t.Fatal(err)
-			} else if !reflect.DeepEqual(names, [][]byte{[]byte("mem")}) {
-				t.Fatalf("unexpected names: %v", BytesToStrings(names))
-			}
-		})
-
-		t.Run("EQREGEX", func(t *testing.T) {
-			names, err := idx.MeasurementNamesByExpr(authorizer, influxql.MustParseExpr(`_name =~ /cpu|disk/`))
-			if err != nil {
-				t.Fatal(err)
-			} else if !reflect.DeepEqual(names, [][]byte{[]byte("cpu")}) {
-				t.Fatalf("unexpected names: %v", BytesToStrings(names))
-			}
-		})
 	})
 }
